+++ proteus/mprans/MCorr.py
@@ -629,6 +629,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.u[0].femSpace.elementMaps.grad_psi,
             self.mesh.nodeArray,
             self.mesh.elementNodesArray,
+            self.elementQuadraturePoints,
             self.elementQuadratureWeights[('u', 0)],
             self.u[0].femSpace.psi,
             self.u[0].femSpace.grad_psi,
@@ -655,6 +656,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.elementDiameter,  # self.mesh.elementDiametersArray,
             self.mesh.nodeDiametersArray,
             self.u[0].dof,
+            self.coefficients.lsModel.u[0].dof,
             self.coefficients.q_u_ls,
             self.coefficients.q_n_ls,
             self.coefficients.ebqe_u_ls,
@@ -744,6 +746,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.u[0].femSpace.elementMaps.grad_psi,
             self.mesh.nodeArray,
             self.mesh.elementNodesArray,
+            self.elementQuadraturePoints,
             self.elementQuadratureWeights[('u', 0)],
             self.u[0].femSpace.psi,
             self.u[0].femSpace.grad_psi,
@@ -768,6 +771,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.elementDiameter,  # self.mesh.elementDiametersArray,
             self.mesh.nodeDiametersArray,
             self.u[0].dof,
+            self.coefficients.lsModel.u[0].dof,
             self.coefficients.q_u_ls,
             self.coefficients.q_n_ls,
             self.coefficients.q_H_vof,
@@ -1030,6 +1034,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.u[0].femSpace.elementMaps.grad_psi,
             self.mesh.nodeArray,
             self.mesh.elementNodesArray,
+            self.elementQuadraturePoints,
             self.elementQuadratureWeights[('u', 0)],
             self.u[0].femSpace.psi,
             self.u[0].femSpace.grad_psi,
@@ -1055,6 +1060,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.elementDiameter,  # self.mesh.elementDiametersArray,
             self.mesh.nodeDiametersArray,
             self.u[0].dof,
+            self.coefficients.lsModel.u[0].dof,
             q_phi,  # self.coefficients.q_u_ls,
             self.coefficients.q_n_ls,
             self.coefficients.ebqe_u_ls,
@@ -1142,6 +1148,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
             self.u[0].femSpace.elementMaps.grad_psi,
             self.mesh.nodeArray,
             self.mesh.elementNodesArray,
+            self.elementQuadraturePoints,
             self.elementQuadratureWeights[('u', 0)],
             self.u[0].femSpace.psi,
             self.u[0].femSpace.grad_psi,
