+++ proteus/tests/levelset/vortex2D/ls_consrv_vortex_2d_n.py
@@ -1,9 +1,12 @@
 from __future__ import absolute_import
 from proteus import *
 from proteus.default_n import *
-from .ls_consrv_vortex_2d_p import *
-from .vortex2D import *
-
+try:
+    from .ls_consrv_vortex_2d_p import *
+    from .vortex2D import *
+except:
+    from ls_consrv_vortex_2d_p import *
+    from vortex2D import *
 
 timeIntegrator = ForwardIntegrator
 timeIntegration = NoIntegration
@@ -87,4 +90,4 @@ else:
 
 conservativeFlux = {}
 if checkMass:
-    auxiliaryVariables = [AuxiliaryVariables.ConservationHistoryMC("vortex2d"+repr(lRefinement)+"p"+repr(pDegree_ls))]
+    pass#auxiliaryVariables = [AuxiliaryVariables.ConservationHistoryMC("vortex2d"+repr(lRefinement)+"p"+repr(pDegree_ls))]
