+++ proteus/tests/levelset/vortex2D/ls_consrv_vortex_2d_p.py
@@ -2,7 +2,10 @@ from __future__ import absolute_import
 from builtins import object
 from proteus import *
 from proteus.default_p import *
-from .vortex2D import *
+try:
+    from .vortex2D import *
+except:
+    from vortex2D import *
 name=soname+"_phicor"
 from proteus.mprans import MCorr
