+++ proteus/tests/levelset/vortex2D/vof_vortex_2d_n.py
@@ -1,8 +1,12 @@
 from __future__ import absolute_import
 from proteus import *
 from proteus.default_n import *
-from .vof_vortex_2d_p import *
-from .vortex2D import *
+try:
+    from .vof_vortex_2d_p import *
+    from .vortex2D import *
+except:
+    from vof_vortex_2d_p import *
+    from vortex2D import *
 
 multilevelNonlinearSolver  = NLNI
 levelNonlinearSolver = Newton
